removes WithPrereqs and BuilderScriptAbstract now takes single generic type
