+++ core/applib/src/test/java/org/apache/isis/testing/fixtures/applib/personas/fixtures/PersonBuilderScript.java
@@ -10,7 +10,7 @@ import lombok.RequiredArgsConstructor;
 
 
 @RequiredArgsConstructor
-public class PersonBuilderScript extends BuilderScriptWithResult<Person, PersonBuilderScript> {
+public class PersonBuilderScript extends BuilderScriptWithResult<Person> {
 
     private final Person_persona persona;
